+++ packages/components/components/elvis-datepicker/src/react/calendar/dateRangeStyles.tsx
@@ -1,4 +1,4 @@
-import { getColor } from '@elvia/elvis-colors';
+import { getThemeColor } from '@elvia/elvis-colors';
 import styled, { css } from 'styled-components';
 import { DayButton } from './calendarStyles';
 
@@ -41,7 +41,7 @@ const getDateRangeBackground = (piece: Partial<DateRangePiece>) => {
       top: 4px;
       left: ${piece.isStart ? '4px' : '0px'};
       border-radius: ${getBorderRadius(piece)};
-      background-color: ${getColor('grey-05')};
+      background-color: ${getThemeColor('state-hover-grey')};
     }
   `;
 };
@@ -73,11 +73,11 @@ export const DateRangeDayContainer = styled.div<Partial<Props>>`
         cursor: pointer;
 
         &:hover ${DayButton} {
-          border-color: ${getColor('elvia-charge')};
+          border-color: ${getThemeColor('state-hover-green')};
         }
 
         ${DayButton} {
-          border-color: ${props.isFocused ? getColor('elvia-charge') : 'transparent'};
+          border-color: ${props.isFocused ? getThemeColor('state-on') : 'transparent'};
         }
       `;
     }
@@ -103,7 +103,7 @@ export const DateRangeDayContainer = styled.div<Partial<Props>>`
     props.isOtherSelectedDate &&
     css`
       ${DayButton} {
-        border-color: ${getColor('elvia-off')};
+        border-color: ${getThemeColor('text-primary')};
       }
     `};
 `;