+++ packages/components/components/elvis-radio-filter/src/react/styledComponents.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { getColor } from '@elvia/elvis-colors';
+import { getThemeColor } from '@elvia/elvis-colors';
 
 export const RadioFilterGroup = styled.div`
   display: flex;
@@ -20,16 +20,13 @@ export const RadioFilterLabel = styled.label<RadioFilterLabelProps>`
   place-items: center;
   border-radius: 32px;
   padding: 4px 12px;
-  border: ${({ isSelected }) => `1px solid ${isSelected ? getColor('black') : 'transparent'}`};
+  border: ${({ isSelected }) => `1px solid ${isSelected ? getThemeColor('text-primary') : 'transparent'}`};
   cursor: pointer;
   position: relative;
 
   &:hover:not(:disabled) {
-    border-color: ${({ isSelected }) => `${isSelected ? getColor('black') : getColor('elvia-charge')}`};
-  }
-  &:focus-within {
-    outline: 2px solid ${getColor('focus-outline')};
-    outline-offset: 2px;
+    border-color: ${({ isSelected }) =>
+      `${isSelected ? getThemeColor('text-primary') : getThemeColor('state-hover-green')}`};
   }
   white-space: nowrap;
   margin: 0;
@@ -53,4 +50,5 @@ export const RadioFilterTitle = styled.span`
   font-weight: 500;
   font-size: 16px;
   line-height: 22px;
+  color: ${getThemeColor('text-primary')};
 `;