+++ packages/components/components/elvis-toolbox/src/index.ts
@@ -32,6 +32,7 @@ export { useIsOverflowing } from './hooks/useIsOverflowing';
 export { useWebComponentState } from './hooks/useWebComponentState';
 export { useSlot } from './hooks/useSlot';
 export { useRovingFocus } from './hooks/useRovingFocus';
+export { useCurrentTheme } from './hooks/useCurrentTheme';
 
 export { BaseProps } from './prop-blocks/baseProps';
 export { ErrorOptions, HasError } from './prop-blocks/hasError';