+++ packages/elvis-assets-icons/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Elvis-assets-icon
 
+## 2.8.0 (01.02.23)
+
+### New feature
+
+- Added new icon `piggy_bank-color`.
+
+## 2.7.1 (30.01.23)
+
+### Patch
+
+- Added support for using css-variables to set the color of icons.
+- Bumped dependency on `@elvia/elvis-colors` to 1.5.0.
+
 ## 2.7.0 (04.01.23)
 
 ### New feature